improve SKIP mechanism visibility

+++ utils/create_template_and_input_data.py
@@ -12,25 +12,32 @@ def create_directory(full_directory_path):
         os.makedirs(full_directory_path)
         print(f"Directory '{full_directory_path}' created successfully.")
     else:
-        print(f"Directory '{full_directory_path}' already exists.")
+        print(f"Directory '{full_directory_path}' already exists - SKIP")
 
 
 def create_file_from_template(directory_path: str, full_directory_path: str, day: str) -> None:
     template = directory_path + "/template/dayXX.py"
     destination = full_directory_path + f"/day{day}.py"
-    shutil.copy(template, destination)
-    print("Template copied to '" + destination + "")
+    if not os.path.exists(destination):
+        shutil.copy(template, destination)
+        print("Template copied to '" + destination + "")
+    else:
+        print("Template not copied - File already exists - SKIP")
 
 
 def create_input_data(directory_path: str, year: str, day: str) -> None:
-    data = get_data(day=int(day), year=int(year))
-    if data:
-        file_path = os.path.join(directory_path, "input.txt")
-        with open(file_path, 'w') as file:
-            file.write(data)
-        print("Data successfully created.")
+    input_file = directory_path + "/input.txt"
+    if not os.path.exists(input_file):
+        data = get_data(day=int(day), year=int(year))
+        if data:
+            file_path = os.path.join(directory_path, "input.txt")
+            with open(file_path, 'w') as file:
+                file.write(data)
+            print("Data successfully created.")
+        else:
+            print("No data found")
     else:
-        print("No data found")
+        print("Input file already exists - SKIP")
 
 
 def create_data(directory_path: str, year: str, day: str) -> None:
